Use stage3 instead of WithBootstrap in distro.
The final Libc pkg in distro needs to use an unpatched version of the tmp
system, which previously was implemented by providing a distro object to the
libc build modified via WithBootstrap. This resulted in strange behavior where
the deps of libc (linux-headers) would actually get built twice, once with the
patched tmp system and once with the unpatched.

This change simplifies things by just creating a separate stage3 that builds
the final libc and linux headers pkgs with the unpatched tmp. It's then
embedded in the distro struct, which uses a pkger that has the patched tmp
system.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>