Cleanup overlay short-name symlinks.
These symlinks are used to minimize the length of the "lowerdir=" args provided
to mount commands, which otherwise was sometimes hitting the 4KB limit on the
size of the data arg provided to the mount syscall. They, however, were not
getting cleaned up before this change, causing re-running the same container
multiple times to fail.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>