Refactor ctr package.
The ctr package previously was mostly remnants of older, ultra-hacky prototype
code that gave me anxiety whenever I needed to make a change to it because it
was so delicate and convoluted. This change cleans up the code by introducing
new interfaces for managing containers/mounts that are clearer and easier
to use. There's still some gnarly portions of the code (setting up the overlay
private dirs and symlink shorteners coming to mind), but they are fairly
concentrated rather than spread all over the place.

I also decided to get rid of a few experimental features that may be nice to
have in the future but are very "extra" and not worth maintaining for the time
being, namely the ability detach and later re-attach to container TTYs and the
ability to create an oci-image and export it to a docker registry. I'll add
those back sometime in the future once more fundamental issues have been solved.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>