Add golang build.
This was commented out previously because it was use a versioning feature that
I had not ported up to the current state of the code. However, I decided that
versioning feature is too complicated and unhelpful, so I just ported the build
of golang to use a more straightforward approach of defining the golang
bootstrap (which is go v1.4) as it's own source package that the current source
package (1.14 at the moment) has a runtime dep on.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>