Remove versioning feature.
This was originally added to support there being different versions of a given
package, which was used at the time to support there being golang-1.4 (which
is used to bootstrap builds of newer versions) and the latest golang version.
However, the feature wasn't great in that it was something enforced at runtime
(not compile time), was complicated to maintain and had simpler alternatives
(namely, just create different package builds for different versions).
Originally, I had thought that maintaining different pkg builds for different
versions of the same thing would be too verbose, but in practice it's very
rarely needed.

There might be a desire for something like this in the future if people create
distros where pkg builds are sourced from many disparate repos and want a way
of asserting that a particular build is at least some version. However, that's
a long ways away (if ever) and should be re-evaluated at that time. Finding
something that can be enforced at compile time would be preferable.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>