Refactor pkg builds.
Pkgs are defined+exposed via functions, which is good because it allows you to
use interfaces to define pkg requirements for an object. However, retrieving
pkgs via functions (as opposed to static fields in a struct) means that the
function needs to return a cached value. If not, it will have to compute the
pkg's state every time it's called. This becomes literally exponentially worse
due to the fact that pkg builds reference other pkgs (their dependencies). So,
without caching, computing a single pkg will result in the entire graph of
dependencies being recursively calculated but with no caching. It's like doing a
depth-first walk but at each vertex of the graph you recursively initiate
another depth-first walk at the vertex before continuing to the next one.
Basically, without caching, a graph of reasonably small size will result in
multiple GBs of memory usage take minutes of CPU time to resolve.

Previously the problem was solved by a combination of the graph.PkgCache
interface and having each pkg define a helper function for accessing the build
of a pkg by checking the cache first.

The previous solution worked but was sort of surprising and clunky to use. If
you want a distro that built pkg Foo, you would ask for a foo.Pkger (which
defines method "Foo() Pkg"). However, in order to actually access the pkg, you
can't just call .Foo(), you have to use the separate helper function,
"foo.Pkg(<foo.Pkger>)". This makes sense only if you know all the inner
machinations of the code, which sucks for an end user.

Additionally, the previous solution made it so that the only way you could
follow this pattern was by defining a totally separate golang package for each
pkg and writing that helper function for accessing it. While that code would
eventually be written by go generate, it still introduced a lot of complication
for use cases where you don't want to "formally" create a new type of pkg and
just want to add something to the distro quickly and easily (for testing, a
one-off pkg custom to your distro build, etc).

The change here gets rid of the need to use the helper function, you can now
just access pkgs directly with the method their Pkger interface defines.
The caching now takes place in pkg build definitions. The main downside is the
builds a little more verbose due to needing to wrap the definition in
"func() Pkg {...}", but is overall easier to understand and less surprising.
For the "one-off" pkg use-case mentioned above, you can either choose to just
not cache (which is probably fine as long as most of the other pkgs in the graph
are using caching) or just write one-line boilerplate for wrapping the build in
"func() Pkg {...}" if caching is really crucial. The main benefit now is that
you can just make that decision when implementing the build rather than when you
access the pkg else where. I feel the trade-off between simplicity in the API is
worth the slight increase in boilerpate (which may become less of a problem if
go generate is used in the future).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>