Fix fifo perms; handle io copy errors.
I tested bincastle on a different environment where I didn't have all caps
in the current userns and got permission denied errors opening fifos. This
is due to the fact that they need to be 0600, but I previously had cap dac
override.

This change also includes the error handling/logging I needed to add in order
to diagnose the issue.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>