Add "AlwaysRun" opt; refactor source opts.
AlwaysRun is a new opt for src pkgs that results in the source being checked and
obtained every invocation of the binary, which is useful for git packages that
work off branches being actively developed on.

The bulk of the change is a refactorization of how options are passed to src
pkgs and setting up similar changes to some other code in the future. The new
way allows helper funcs like src.Curl and src.Git define exactly the args they
accept while still allowing args to be required or optional. It also uses a
generic backend under util that other parts of the code (such as the pkg code)
will likely use in the future too.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>