SIGTERM containers instead of SIGKILL.
SIGKILL was previously being used to ensure a container was absolutely dead when
shutting down. However, this meant that the container receving SIGKILL wouldn't
always run all of its cleanup before existing, leaving around state to be
removed manually next run. SIGTERM doesn't have that problem (as the signal
handler is overrode in AttachConsole) and seems to work fine for the moment.
Eventually, there should be a timeout with a fallback to SIGKILL if SIGTERM
doesn't work.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>