Run final interactive container via buildkit solve.
Previously, buildkit solves were used to build containers, but the final one
presented to the users as an interactive shell was created separately. This
created lots of complication, extra code and problems due to the need to
export all of the images created during the build and create a construct the
final container out of those. The biggest bug resulting was that garbage
collection of state by the buildkit daemon no longer functioned correctly; even
if it was disabled entirely buildkit internally would panic when it found that
databases it expected to be opened for tracking state usage had been closed.

The new approach updates the custom buildkit worker to support an interactive
mode where a container's io is attached directly to the process's tty and
never exports any of the file system changes to buildkit's cache.

This new approach had previously been avoided because it felt hacky, but in
retrospect the alternative ended up much worse. This solution is much simpler
in that it allows buildkit to take care of all the details of managing container
fs state/cache.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>