Split container cleanup stack into pre and post kill.
This allows the container's pty to be closed before the rest of the kill/cleanup
process proceeds, which in turn leads to better behavior when killing processes
that respond "better" to a SIGHUP than a SIGTERM (i.e. bash, which ignores
SIGTERMs).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

return correct error value in buildkit worker

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

fixup! return correct error value in buildkit worker