Dynamically obtain llb from git (via llbbuild).
Previously, the code llb defining the system you want to run was just embedded
in the same binary as the rootless buildkit daemon. This meant that you needed
to rebuild that binary every time you changed even just the system definition,
in addition to essentially limiting you to writing that system definition in
the Go DSL defined in this repo.

This change uses the llbbuild interface provided by buildkit to instead
obtain the LLB for the system at binary-runtime. The command line args have
been updated to ask for a git url, git ref and a subpath within that git repo
to a Go binary that outputs LLB to stdout when executed. Via llbbuild, a
container is spawned that builds and obtains that llb, which is then provided
to the solver and ran just as before.

This is great for a bunch of reasons:
* It offers a clear path for eventually supporting system specifications created
  in different languages (the Go one that exists will be decent after some more
  cleanup, but it would be unfortunate to only ever support Go).
* It even offers a potential path for making bincastle a more generic tool that
  can build+execute completely arbitrary LLB (whereas right now there are a lot
  of ties between the LLB created by the Go DSL and the custom worker bincastle
  uses at the moment).
  * This path becomes more clear if some of the hacks in the custom worker are
    merged upstream to buildkit (such as being able to merge mounts and "exec"
    containers that don't export to cache and can be hooked up directly to a
    user's pty).
* It makes the development loop easier as you don't need to rebuild a binary
  if you are only changing the system definition (i.e. the code that generates
  LLB).

I also looked into just making bincastle use the "gateway" frontend, which
allows you to create a container image containing a binary that issues solve
requests back to the main buildkit daemon. However, this ended up being more
complicated than using llbbuild because bincastle, by it's nature, already
asks users to download a binary that they will use to execute their systems.
Defining a separate image containing another binary that issues solves requests
wouldn't have provided benefit on top of just putting the llbbuild logic
directly in the bincastle binary itself.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>