Add basic support for remote cache/import.
This allows you to use buildkit's support for exporting/importing cache to/from
a registry. The biggest benefit is that if you start bincastle on a new machine
you can just import cache from a previous build rather than being forced to
always rebuild the entire unchanged system from scratch.

The change also upgrades buildkit to the latest commit in order to pick up a
fix for a bug in the remotecache code that caused the cache import to be
unusably slow.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>