Use Replace instead of Unbootstrap.
Unbootstrap was a bit hard to understand and somewhat surprising at times,
specifically due to the fact that whether a dependency was still present after
the unbootstrap came down to whether it was still "reachable", which in turn
depended on the computed transitive reduction.

Replace satisfies the same needs that led to unbootstrap, but in more
straightforward way; it allows you to simply specify that one spec should be
replaced by another (or replaced by nil if you want that spec to be removed).
The only downside is this more generic power allows you to create loops, but
for now that's an acceptable tradeoff.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>