Add fuse-overlayfs target.
This is the first step for supporting use of fuse-overlayfs instead of native
overlayfs in bincastle (which is desired because native overlayfs only works
inside an unpriv userns w/ a kernel patch like Ubuntu has).

The new Makefile target will build fuse-overlayfs inside a bincastle system
and then export a single layer image consisting of just "/fuse-overlayfs".

The next steps after this change will be to have bincastle check if it needs
fuse-overlayfs on start, and, if so, get the fuse-overlayfs binary from the
a specified image (presumed to be one exported by a previous run of the new
Makefile target).

There are trade-offs of this approach relative to alternatives like embedding
fuse-overlayfs as library or embedding the fuse-overlayfs binary in bincastle,
but those will be detailed in future commits.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>