Add support for optionally using fuse-overlayfs.
Some distros like Ubuntu come by default with a kernel patch that enables
overlay mounts to be created from within unprivileged user namespaces. However,
on other platforms, overlays will fail inside bincastle. This change enables
use of fuse-overlayfs on those platforms, which only requires a kernel of
version 4.18 or greater.

The implementation right now relies on pulling down a fuse-overlayfs binary
when needed at runtime. One preferable alternative would have been to just
build fuse-overlayfs code inside bincastle and use it as a library. However,
that would require some upstream modifications of the fuse-overlayfs, be tricky
to write as it would involve CGo and also introduce licensing complications due
to the fact that fuse-overlayfs uses a GPL license, which restricts linking
against its code unless bincastle also switched to GPL. There may be ways to
work around the licensing and technical hurdles, but for now its not worth it.

Another possibility is to embed an already built fuse-overlayfs binary inside
the bincastle binary, which probably wouldn't violate GPL as it should count as
"mere aggregation". However, IANAL and that approach is fairly complicated from
a technical perspective anyways, so it's not being pursued now either.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>