Initial migration to frontend-based implementation.
The primary change here is to register a frontend for bincastle and migrate
a few parts of the logic surrounding exec containers from the executor to this
new frontend. The most immediate benefit is that the upperdir of execs are now
managed as an actual mutable ref in the frontend, which is a lot less hacky
than the previous approach where there was just a random dir that got saved
across execs and re-used (though it's still hacky; baby steps).

The frontend is also the better long-term direction anyways as it is one step
closer to being able to use completely vanilla buildkit code. Eventually, the
desire is to just have a gateway-based frontend, but there will need to be a way
to manage mutable refs before that's possible.

A bunch of other changes came along with this (some only tangentially related):
* The container setup is modified a bit; previously there was an initial outer
  container created that ran both buildkitd and the buildkit client. Now, thanks
  mostly to the frontend, that outer container only needs to run buildkitd; the
  client can remain in the non-containerized process. This has the immediate
  benefit of simplifying support for things like local dirs and longer-term
  benefits for my plans to support multiple bincastle execs and switching
  between those execs dynamically.
* Removed the (real) overlayfs backend; now only fuse-based overlays are
  supported. This is due to a long-standing bug I finally encountered involving
  opaque directories; specifically, when a directory is created in an upperdir
  and there is no corresponding dir in a lowerdir, the directory is set with
  the "opaque" xattr, meaning reads on it will stop recursing to lower dirs when
  used as part of an overlay. This is fine while it's an upperdir, but once it
  moves to be a lowerdir in a later mount, it can cause unexpected behavior. You
  can solve this problem by just making directories not-opaque after they're
  done being an upperdir, but turns out doing that is impossible with the
  overlays enabled inside unpriv userns because while the kernel patches allow
  you to create overlays, they don't allow you to modify trusted xattrs created
  by those overlays. Fuse-overlayfs on the hand uses untrusted xattrs which can
  be modified, so it's the only implementation feasible right now.
* Bincastle graphs now have their own serialization format, which is how they
  get sent to the frontend. This code sucks but is also just a holdover until
  there is an official merge-op in buildkit, at which time it can be made to
  suck less.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>