First-pass support for dynamically switching bincastle execs.
This is a basic rough implementation of support for the ability to within one
bincastle exec talk back to the bincastle daemon, request a new exec, have
the pty switch to that one and then switch back to the original after the new
exec exits. This allows you to test new system builds from within an existing
one, with the end vision being bincastle serving as a sort of system build
playground. It will become even more useful and interesting once support for
local source overrides are added in a near-future commit.

There are a ton of caveats to this right now:
* Your upper dir will always transparently follow you to the new exec. This
  is the desired behavior in some situations, but it may result in unexpected
  behavior if the new system's fs is radically different than the previous
  one's. In the long-term this will be improved by adding a lot more options
  for managing upperdirs (likely along-side some sort of ncurses-like
  interface).
* When you switch back to the previous system, while your upperdir will transfer
  transparently, all the previous processes will have been killed. This is
  needed at the moment because upperdir transfer between execs, meaning the
  overlays need to be torn down and reset up. In the long term, this may be able
  to be alleviated with a combination of better upperdir management (so you can
  optionally "fork" an upperdir to a new ephemeral copy without tearing it down)
  and cgroup freezer stuff (or perhaps even unprivileged CRIU someday).
* The ability to switch back to previous systems has a "depth" of 1; that is to
  say you can only switch back to the immediate previous system. After that,
  exiting the shell will just result in bincastle exiting.
* You can only run one exec at a time.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>