Add full README and demo.
Signed-off-by: Erik Sipsma <erik@sipsma.dev>

Move distro back to this repo.

I had previously moved the example distro outside the bincastle repo to its own
repo, partially because I wanted to make it clear that "distro != bincastle",
which is to say that the distro is an example of using bincastle rather than
what bincastle inherently is.

However, at the moment that ideal of being complete detached from bincastle
is not fully achieved (distro is currently used to build fuse-overlayfs, for
one example) and managing it in a separate repo became extremely annoying with
little benefit.

Now, distro is back in this repo but under the examples directory.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

Add support for TTY-based build progress.

This change is complicated by the fact that the client needs to know when the
build is done so it can stop displaying the progress and let the exec take over
the TTY. This required splitting exec into 2 steps, the pre-build and the actual
exec. The TTY progress is only enabled for the pre-build, which takes care of
all the actual work.

You can get the old progress output using the --verbose flag.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

Re-add support for outputting systemd def in json format.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

Generate more consistent graph digests.

This ensures that equivalent graphs re-used the same cache as much as possible.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

Remove some host-specific assumptions.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>